feat(dtype): support JAX x64 mode by removing float32 hardcoding
Every TFP-backed Distribution silently downcast its parameters to
float32 in __init__, which made the library incompatible with
jax.config.update("jax_enable_x64", True) — log_prob/sample/mean raised
TypeError when given float64 input.

ProbPipe now follows JAX's dtype rules: the default float dtype tracks
jnp.zeros(()).dtype (float32 normally, float64 with x64 enabled), and
explicit user dtypes are preserved. Multi-parameter distributions
promote inputs to a common float dtype via jnp.result_type; integer
inputs are promoted to JAX's default float so e.g. Normal(loc=0,
scale=1) still works.

A new probpipe/_dtype.py exposes two helpers (_default_float_dtype and
_promote_floats) used across distributions, weights, empirical classes,
bootstrap, joint distributions, GRF, KDE, broadcast outputs, GLM, Stan,
and TFP-MCMC init state. Weights.dtype and the base
NumericRecordDistribution.dtypes now reflect actual data dtype instead
of returning a hardcoded float32 literal.

The float64-truncation warning filter in probpipe/__init__.py is
removed, since the warning should no longer fire under correct usage.

Test updates: a new tests/test_dtype.py covers x32, x64, and mixed-dtype
paths via subprocess isolation; existing dtype == jnp.float32 assertions
are made dtype-agnostic; loc fixtures use JAX's default float so they
share dtype with their cov_matrix counterparts under x64;
tfb.Shift(5.0)-style bijector constants are wrapped in jnp.array so
they honor x64.

Verified: 2113 passing under x32, 2102 passing under x64.

+++ probpipe/core/_broadcast_distributions.py
@@ -380,7 +380,7 @@ def _make_marginal(
                 pass
         try:
             stacked = jnp.stack(
-                [jnp.asarray(r, dtype=jnp.float32) for r in output_samples], axis=0
+                [jnp.asarray(r) for r in output_samples], axis=0
             )
             return _ArrayMarginal(stacked, weights, name=name)
         except (ValueError, TypeError):
@@ -391,7 +391,7 @@ def _make_marginal(
         return _ListMarginal(output_samples, weights, name=name)
 
     # Single array result (e.g., from vmap); ensure at least 1D for the sample axis
-    arr = jnp.atleast_1d(jnp.asarray(output_samples, dtype=jnp.float32))
+    arr = jnp.atleast_1d(jnp.asarray(output_samples))
     return _ArrayMarginal(arr, weights, name=name)
